Introduce `Hasher.putShort()`
So the build cache can distinguish between short and int inputs.

+++ platforms/core-configuration/model-core/src/main/java/org/gradle/internal/snapshot/impl/ShortValueSnapshot.java
@@ -25,6 +25,6 @@ public class ShortValueSnapshot extends AbstractIsolatableScalarValue<Short> {
 
     @Override
     public void appendToHasher(Hasher hasher) {
-        hasher.putInt(getValue());
+        hasher.putShort(getValue());
     }
 }
